C OpenGL Projects
This repository contains some of my OpenGL projects done in C language.
They were all for completely academic purpose.

In these projects we tried to create a replica of a selected location.
We have the Smrity Shoudho of Bangladesh, Shahid Minar of Bangladesh and
Faisal Mosque of Pakistan. The created 3D objects are not at all similar
with the real ones but we tried to create at least something that look
alike.